requirements.txt
Django
django-db-logger
django-simple-captcha
Pillow
six
djangorestframework
mysqlclient